Organize external API
