+++ matrix-bot/matrix-bot.cabal
@@ -59,11 +59,14 @@ library
     hs-source-dirs:   src
     exposed-modules:  MatrixBot.Main
                     , Network.Matrix.Bot
-    other-modules:    Network.Matrix.Bot.Async
-                    , Network.Matrix.Bot.ErrorHandling
+                    , Network.Matrix.Bot.Async
                     , Network.Matrix.Bot.Event
-                    , Network.Matrix.Bot.JSON
                     , Network.Matrix.Bot.Router
+    other-modules:    Network.Matrix.Bot.Async.Internal
+                    , Network.Matrix.Bot.ErrorHandling
+                    , Network.Matrix.Bot.Event.Internal
+                    , Network.Matrix.Bot.JSON
+                    , Network.Matrix.Bot.Router.Internal
                     , Network.Matrix.Bot.State
                     , Network.Matrix.Bot.Sync
     default-language: Haskell2010
